fastapi-gelai/pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "edgedb>=2.2.0",
    "fastapi[standard]>=0.115.6",
    "openai>=1.59.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
